+++ test/errorBoundary.spec.tsx
@@ -1,7 +1,8 @@
 import _ from 'lodash'
 import React from 'react'
-import { createAppHost, addMockShell, renderInHost, withConsoleErrors } from '../testKit'
+import { createAppHost, addMockShell, renderInHost, withConsoleErrors, isNode } from '../testKit'
 import { ErrorBoundary } from '../src'
+import { act } from 'react-test-renderer'
 
 describe('ErrorBoundary', () => {
     let renderShouldThrow: boolean
@@ -26,7 +27,7 @@ describe('ErrorBoundary', () => {
             shell
         )
 
-        expect(root?.exists('.test-comp')).toBe(true)
+        expect(root.root.findAll(x => x.props.className === 'test-comp').length).toBe(1)
     })
 
     it('should render empty on error', () => {
@@ -42,8 +43,14 @@ describe('ErrorBoundary', () => {
             shell
         )
 
-        expect(root?.exists('.test-comp')).toBe(false)
-        expect(root?.isEmptyRender()).toBe(true)
+        const node = root.toJSON()
+
+        if (!isNode(node)) {
+            fail('Expected node')
+        }
+
+        expect(root.root.findAll(x => x.props.className === 'test-comp').length).toBe(0)
+        expect(node.children).toBeNull()
     })
 
     it('should recover by change in store', () => {
@@ -51,7 +58,7 @@ describe('ErrorBoundary', () => {
         const shell = addMockShell(host)
 
         renderShouldThrow = true
-        const { root } = renderInHost(
+        const { root, rootComponent } = renderInHost(
             <ErrorBoundary shell={shell}>
                 <TestComponent />
             </ErrorBoundary>,
@@ -59,14 +66,16 @@ describe('ErrorBoundary', () => {
             shell
         )
 
-        expect(root?.exists('.test-comp')).toBe(false)
+        expect(root.root.findAll(x => x.props.className === 'test-comp').length).toBe(0)
 
         renderShouldThrow = false
-        host.getStore().dispatch({ type: 'TEST' })
-        host.getStore().flush()
-        root?.update()
+        act(() => {
+            host.getStore().dispatch({ type: 'TEST' })
+            host.getStore().flush()
+            root.update(rootComponent)
+        })
 
-        expect(root?.exists('.test-comp')).toBe(true)
+        expect(root.root.findAll(x => x.props.className === 'test-comp').length).toBe(1)
     })
 
     it('should show sticky error', () => {
@@ -83,11 +92,13 @@ describe('ErrorBoundary', () => {
                 shell
             )
         )
+        expect(root.root.findAll(x => x.props.className === 'test-comp').length).toBe(0)
+        const errors = root.root.findAll(x => x.props.className?.includes('component-error'))
 
-        expect(root?.exists('.test-comp')).toBe(false)
-        expect(root?.exists('.component-error')).toBe(true)
-        expect(root?.find('.component-error').text()).toContain(`error in ${shell.name} / test_comp`)
-        expect(root?.find('.component-error').find('button').text()).toBe('reset')
+        expect(errors.length).toBe(1)
+        expect(errors[0].children[0]).toBe('error in ')
+        expect((errors[0].children[1] as any).children[0]).toBe(`${shell.name} / test_comp`)
+        expect(errors[0].findByType('button').children[0]).toBe('reset')
     })
 
     it('should keep sticky error after change in store', () => {
@@ -95,7 +106,7 @@ describe('ErrorBoundary', () => {
         const shell = addMockShell(host)
 
         renderShouldThrow = true
-        const { root } = withConsoleErrors(() =>
+        const { root, rootComponent } = withConsoleErrors(() =>
             renderInHost(
                 <ErrorBoundary shell={shell} componentName="test_comp">
                     <TestComponent />
@@ -106,14 +117,19 @@ describe('ErrorBoundary', () => {
         )
 
         renderShouldThrow = false
-        host.getStore().dispatch({ type: 'TEST' })
-        host.getStore().flush()
-        root?.update()
+        act(() => {
+            host.getStore().dispatch({ type: 'TEST' })
+            host.getStore().flush()
+            root.update(rootComponent)
+        })
 
-        expect(root?.exists('.test-comp')).toBe(false)
-        expect(root?.exists('.component-error')).toBe(true)
-        expect(root?.find('.component-error').text()).toContain(`error in ${shell.name} / test_comp`)
-        expect(root?.find('.component-error').find('button').text()).toBe('reset')
+        expect(root.root.findAll(x => x.props.className === 'test-comp').length).toBe(0)
+        const errors = root.root.findAll(x => x.props.className?.includes('component-error'))
+
+        expect(errors.length).toBe(1)
+        expect(errors[0].children[0]).toBe('error in ')
+        expect((errors[0].children[1] as any).children[0]).toBe(`${shell.name} / test_comp`)
+        expect(errors[0].findByType('button').children[0]).toBe('reset')
     })
 
     it('should reset sticky error on reset button click', () => {
@@ -121,7 +137,7 @@ describe('ErrorBoundary', () => {
         const shell = addMockShell(host)
 
         renderShouldThrow = true
-        const { root } = withConsoleErrors(() =>
+        const { root, rootComponent } = withConsoleErrors(() =>
             renderInHost(
                 <ErrorBoundary shell={shell} componentName="test_comp">
                     <TestComponent />
@@ -131,14 +147,16 @@ describe('ErrorBoundary', () => {
             )
         )
 
-        const resetButton = root?.find('.component-error').find('button')
-        expect(resetButton?.exists()).toBe(true)
+        const button = root.root.findByType('button')
+        expect(button).toBeDefined()
 
         renderShouldThrow = false
-        resetButton?.simulate('click')
-        root?.update()
+        button.props.onClick()
+        act(() => {
+            root.update(rootComponent)
+        })
 
-        expect(root?.exists('.test-comp')).toBe(true)
-        expect(root?.exists('.component-error')).toBe(false)
+        expect(root.root.findAll(x => x.props.className === 'test-comp').length).toBe(1)
+        expect(root.root.findAll(x => x.props.className?.includes('component-error')).length).toBe(0)
     })
 })
